+++ src/main/java/co/orange/ddanzi/domain/product/Category.java
@@ -1,6 +1,6 @@
 package co.orange.ddanzi.domain.product;
 
-import co.orange.ddanzi.common.domain.BaseTimeEntity;
+import co.orange.ddanzi.global.common.domain.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
