+++ src/main/java/co/orange/ddanzi/global/common/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package co.orange.ddanzi.common.domain;
+package co.orange.ddanzi.global.common.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
