+++ src/main/java/co/orange/ddanzi/global/common/response/ApiResponse.java
@@ -1,6 +1,6 @@
-package co.orange.ddanzi.common.response;
+package co.orange.ddanzi.global.common.response;
 
-import co.orange.ddanzi.common.exception.Error;
+import co.orange.ddanzi.global.common.exception.Error;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.Builder;
